pyproject.toml
[tool.poetry]
name = "colli_parser"
version = "0.1.0"
description = "Parses invoices to JSON using ChatGPT."
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
openai = "^1.0"
pydantic = "^2.0"
loguru = "^0.7.0"
pdf2image = "^1.17.0"
polars = "^1.22.0"
ipykernel = "^6.29.5"
pandas = "^2.2.3"
pyarrow = "^19.0.1"
polars-ds = "^0.8.1"
flask = "^3.1.0"
python-dotenv = "^1.0.1"
python-telegram-bot = "^21.10"
quart = "^0.20.0"
fastapi = "^0.115.8"
uvicorn = "^0.34.0"
splitwise = "^3.0.0"
tabulate = "^0.9.0"
regex = "^2024.11.6"
python-intervals = "^1.10.0.post1"
azure-functions = "^1.21.3"
mistralai = "^1.5.1"

[tool.poetry.dev-dependencies]
